refactor(auth): move reusable auth composables to `AuthComposables.kt`
