+++ feature/authentication/src/main/java/com/crisiscleanup/feature/authentication/ui/AuthenticateScreen.kt
@@ -2,29 +2,15 @@ package com.crisiscleanup.feature.authentication.ui
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.sizeIn
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -34,14 +20,10 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.crisiscleanup.core.designsystem.LocalAppTranslator
@@ -49,14 +31,10 @@ import com.crisiscleanup.core.designsystem.component.BusyButton
 import com.crisiscleanup.core.designsystem.component.OutlinedClearableTextField
 import com.crisiscleanup.core.designsystem.component.OutlinedObfuscatingTextField
 import com.crisiscleanup.core.designsystem.component.actionHeight
-import com.crisiscleanup.core.designsystem.theme.CrisisCleanupTheme
-import com.crisiscleanup.core.designsystem.theme.DayNightPreviews
 import com.crisiscleanup.core.designsystem.theme.LocalFontStyles
-import com.crisiscleanup.core.designsystem.theme.actionLinkColor
 import com.crisiscleanup.core.designsystem.theme.fillWidthPadded
 import com.crisiscleanup.core.designsystem.theme.listItemModifier
 import com.crisiscleanup.core.designsystem.theme.listItemPadding
-import com.crisiscleanup.core.designsystem.theme.primaryBlueColor
 import com.crisiscleanup.core.ui.rememberCloseKeyboard
 import com.crisiscleanup.core.ui.rememberIsKeyboardOpen
 import com.crisiscleanup.core.ui.scrollFlingListener
@@ -65,7 +43,6 @@ import com.crisiscleanup.feature.authentication.AuthenticationViewModel
 import com.crisiscleanup.feature.authentication.BuildConfig
 import com.crisiscleanup.feature.authentication.R
 import com.crisiscleanup.feature.authentication.model.AuthenticationState
-import com.crisiscleanup.core.common.R as commonR
 
 @Composable
 fun AuthRoute(
@@ -158,116 +135,6 @@ private fun AuthenticateScreen(
                 }
             }
         }
-    }
-}
-
-@Composable
-internal fun CrisisCleanupLogoRow() {
-    // TODO Adjust to other screen sizes as necessary
-    Box(Modifier.padding(top = 16.dp, start = 8.dp)) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Start,
-        ) {
-            Image(
-                painterResource(R.drawable.worker_wheelbarrow_world_background),
-                modifier = Modifier
-                    .testTag("ccuBackground")
-                    .padding(top = 32.dp)
-                    .size(width = 480.dp, height = 240.dp)
-                    .offset(x = 64.dp),
-                contentScale = ContentScale.FillHeight,
-                contentDescription = null,
-            )
-        }
-        Row(
-            modifier = fillWidthPadded,
-            horizontalArrangement = Arrangement.Start,
-        ) {
-            Image(
-                modifier = Modifier
-                    .testTag("ccuLogo")
-                    .sizeIn(maxWidth = 160.dp),
-                painter = painterResource(commonR.drawable.crisis_cleanup_logo),
-                contentDescription = stringResource(com.crisiscleanup.core.common.R.string.crisis_cleanup),
-            )
-        }
-    }
-}
-
-@Composable
-fun LoginWithDifferentMethod(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit,
-) {
-    val translator = LocalAppTranslator.current
-    TextButton(
-        modifier = modifier.padding(horizontal = 16.dp),
-        onClick = onClick,
-        shape = RoundedCornerShape(4.dp),
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(
-                Icons.Default.ArrowBack,
-                contentDescription = "Login using different method",
-                tint = actionLinkColor,
-                modifier = Modifier
-                    .padding(horizontal = 4.dp)
-                    .size(24.dp),
-            )
-            Text(
-                text = translator(
-                    "~~Login using different method",
-                    R.string.loginUsingDifferentMethod,
-                ),
-                color = actionLinkColor,
-                style = LocalFontStyles.current.header3,
-                modifier = Modifier.padding(horizontal = 4.dp),
-            )
-        }
-    }
-}
-
-@Composable
-private fun ConditionalErrorMessage(errorMessage: String) {
-    if (errorMessage.isNotEmpty()) {
-        Text(
-            modifier = fillWidthPadded,
-            text = errorMessage,
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.error,
-        )
-    }
-}
-
-@Composable
-private fun LinkAction(
-    textTranslateKey: String,
-    modifier: Modifier = Modifier,
-    arrangement: Arrangement.Horizontal = Arrangement.End,
-    enabled: Boolean = false,
-    action: () -> Unit = {},
-) {
-    val translator = LocalAppTranslator.current
-    Row(
-        Modifier.fillMaxWidth(),
-        horizontalArrangement = arrangement,
-    ) {
-        Text(
-            text = translator(textTranslateKey),
-            modifier = Modifier
-                .clickable(
-                    enabled = enabled,
-                    onClick = action,
-                )
-                .then(modifier),
-            style = LocalFontStyles.current.header4,
-            color = primaryBlueColor,
-        )
     }
 }
 
@@ -437,22 +304,4 @@ internal fun AuthenticatedScreen(
         enabled = isNotBusy,
         action = closeAuthentication,
     )
-}
-
-@DayNightPreviews
-@Composable
-fun LogoRowPreview() {
-    CrisisCleanupTheme {
-        CrisisCleanupLogoRow()
-    }
-}
-
-@DayNightPreviews
-@Composable
-fun LoginWithDifferentMethodPreview() {
-    CrisisCleanupTheme {
-        LoginWithDifferentMethod(
-            onClick = {},
-        )
-    }
 }
